Split query and mutation objects into different
