+++ workspace-service/src/graphql/mod.rs
@@ -1,0 +1,45 @@
+mod folders;
+mod workspaces;
+
+use super::db;
+use async_graphql::{
+    http::{playground_source, GraphQLPlaygroundConfig},
+    EmptySubscription, GQLMergedObject, Schema,
+};
+use sqlx::PgPool;
+use tide::{http::mime, Request, Response, StatusCode};
+
+#[derive(Clone)]
+pub struct State {
+    schema: Schema<Query, Mutation, EmptySubscription>,
+}
+
+impl State {
+    pub fn new(pool: PgPool) -> State {
+        State {
+            schema: Schema::build(Query::default(), Mutation::default(), EmptySubscription)
+                .data(pool)
+                .finish(),
+        }
+    }
+}
+
+#[derive(GQLMergedObject, Default)]
+struct Query(folders::FoldersQuery, workspaces::WorkspacesQuery);
+
+#[derive(GQLMergedObject, Default)]
+struct Mutation(folders::FoldersMutation, workspaces::WorkspacesMutation);
+
+pub async fn handle_graphql(req: Request<State>) -> tide::Result {
+    let schema = req.state().schema.clone();
+    async_graphql_tide::graphql(req, schema, |query_builder| query_builder).await
+}
+
+pub async fn handle_graphiql(_: Request<State>) -> tide::Result {
+    let response = Response::builder(StatusCode::Ok)
+        .body(playground_source(GraphQLPlaygroundConfig::new("/graphql")))
+        .content_type(mime::HTML)
+        .build();
+
+    Ok(response)
+}
